requirements.txt
Django==4.0
djangorestframework==3.13.1
Pillow
factory-boy==3.2.1
drf-yasg
psycopg2
tzdata